schemas: make `direction` in `listpeerchannels` a u32
everywhere else it's a u32 aswell, an `integer` makes it a i64
but possible values are only 0 and 1

Changelog-None

+++ cln-rpc/src/model.rs
@@ -7555,7 +7555,7 @@ pub mod responses {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub closer: Option<ChannelSide>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub direction: Option<i64>,
+	    pub direction: Option<u32>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub dust_limit_msat: Option<Amount>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
